The Big Fix :new_moon_with_face:

+++ Algorithms/Greedy-Algorithms/largest_number.cpp
@@ -8,22 +8,22 @@
 #include <vector>
 #include <string>
 
-std::string LargestNumber(const size_t& numDigits) {
-    std::vector<size_t> digitCounter(10);
-    for (size_t i = 0; i < numDigits; ++i) {
-        size_t digit;
+std::string LargestNumber(const unsigned& numDigits) {
+    std::vector<unsigned> digitCounter(10);
+    for (unsigned i = 0; i < numDigits; ++i) {
+        unsigned digit;
         std::cin >> digit;
         ++digitCounter[digit];
     }
     std::string largest;
-    for (size_t i = 9; i >= 0; --i) {
+    for (int i = 9; i >= 0; --i) {
         largest += std::string(digitCounter[i], (char) (i + 48));
     }
     return largest;
 }
 
 int main() {
-    size_t numDigits;
+    unsigned numDigits;
     std::cin >> numDigits;
     std::cout << LargestNumber(numDigits) << std::endl;
